Refactoring to make things easier to read
