+++ bot/normal/helpers.go
@@ -1,0 +1,51 @@
+package normal
+
+import (
+	"encoding/hex"
+	"fmt"
+	"time"
+
+	"github.com/pkg/errors"
+)
+
+func hexToRaw(hexBytes []byte) ([]byte, error) {
+	raw := make([]byte, hex.DecodedLen(len(hexBytes)))
+	n, err := hex.Decode(raw, hexBytes)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to decode hex")
+	}
+	if n != len(raw) {
+		return nil, fmt.Errorf("failed to decode hex: decoded %d bytes, expected to decode %d bytes", n, len(raw))
+	}
+	return raw, nil
+}
+
+func max(a, b uint64) uint64 {
+	if a > b {
+		return a
+	}
+	return b
+}
+
+func abs(a int64) int64 {
+	if a < 0 {
+		return -a
+	}
+	return a
+}
+
+func doze(d time.Duration, stop chan bool) error {
+	interval := 100 * time.Millisecond
+	for d > interval {
+		select {
+		case <-stop:
+			return errors.New("doze interrupted")
+
+		default:
+			time.Sleep(interval)
+			d -= interval
+		}
+	}
+	time.Sleep(d)
+	return nil
+}
